UI/requirements.txt
streamlit
pillow
numpy
langdetect
requests
pandas
matplotlib
